Fix some of the compiler warnings.

+++ modules/spaint/include/spaint/core/SpaintPipeline.h
@@ -85,7 +85,7 @@ private:
   LowLevelEngine_Ptr m_lowLevelEngine;
 
   /** The maximum number of voxels to sample for each label. */
-  int m_maxVoxelsPerLabel;
+  size_t m_maxVoxelsPerLabel;
 
   /** The mode in which the pipeline is currently running. */
   Mode m_mode;
